+++ src/protocol/dap.cpp
@@ -729,6 +729,45 @@ HRESULT DAP::HandleCommand(const std::string &command, const nlohmann::json &arg
                 responseBody.emplace("totalModules", totalModules);
 
                 return S_OK;
+            }},
+        {"gotoTargets", [&](const json &arguments, json &responseBody)
+            {
+                const auto &sourceJson = arguments.at("source");
+                const std::string sourcePath = sourceJson.value("path", std::string());
+                const std::string sourceName = sourceJson.value("name", std::string());
+                if (sourcePath.empty() && sourceName.empty())
+                {
+                    return E_INVALIDARG;
+                }
+                Source source(sourcePath.empty() ? sourceName : sourcePath);
+                if (sourceJson.contains("checksums"))
+                {
+                    std::transform(sourceJson.at("checksums").cbegin(), sourceJson.at("checksums").cend(),
+                                   std::back_inserter(source.checksums), [](const auto &c)
+                                   {
+                                       return Checksum(c.value("algorithm", std::string()),
+                                                       c.value("checksum", std::string()));
+                                   });
+                }
+
+                const int32_t line = arguments.at("line");
+                const int32_t column = arguments.value("column", 0);
+
+                HRESULT Status = S_OK;
+                std::vector<GotoTarget> targets;
+                std::string output;
+                if (FAILED(Status = m_sharedDebugger->GetGotoTarget(source, line, column, targets, output)))
+                {
+                    if (!output.empty())
+                    {
+                        responseBody.emplace("message", output);
+                    }
+                    return Status;
+                }
+
+                responseBody.emplace("targets", targets);
+
+                return S_OK;
             }}};
 
     if (m_sharedDebugger == nullptr)